Write the generated document even when generation reports errors
Loading a real application produced a 5,362-line document that compiles
cleanly under the Screenplay compiler, alongside error diagnostics about
constructs the generator could not read. The command threw all of it
away and left the user with nothing at all — the one outcome that helps
nobody. A document that is very nearly right, plus honest diagnostics
saying where it is not, is strictly more useful.

With --file the document is now always written and the command still
exits non-zero, so a CI gate keeps failing while a person gets something
to look at. Writing to standard output is unchanged: whatever consumes
'screenplay generate > MyApp.play' cannot tell a partial document from a
complete one, so nothing is written there.
